Separate code into dedicated modules
