+++ modules/globals.py
@@ -1,0 +1,1 @@
+FULL_SETUP_URL = 'https://github.com/OhItsTom/Spotify-Version/releases/download/spotify-1-1-62-583/spotify-1-1-62-583.exe'
